feat: add lesson and attend auth

+++ src/main/java/com/lms/notification/NotificationResponse.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class NotificationResponse {
-    int userId;
+    Long userId;
     int notificationId;
     String message;
     boolean isRead;
